error API fixed

+++ backend/src/main/java/com/teacher/manager/dto/SubjectDTO.java
@@ -11,7 +11,7 @@ public class SubjectDTO {
   private Long id;
   private String code;
   private String name;
-  private int credits;
+  private double credits;
   private Long groupId;
   private String groupName;
 }
